backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pytesseract
Pillow
